remove pre twitter api and fully replace with scraper lib
